Impemented restricted-scope interests in ccnd.  This uses a new optional Scope element Interest messages.  Scope 0 interests will only go as far as the ccnd and my be used to get content out of the content store without bothering any other clients.  Scope 1 is restricted to the local machine (whatever that means these days).  Larger values are currently unspecified.
The c client interface was extended to allow things other than the Name to be expressed in the interest by providing a template ccnb message.  This gets you the ability to send PublisherID and Scope and whatever else we think of without rewhacking the client interface.

ccntimestamps now uses scope 0, so you get to see the timestamps of things in the content store without provoking any changes.

Refactored ccn_buf_decoder.c to remove somed cloned code for parsing PublisherID.

Added a -reconnect option to dataresponsetest so it can survive a restart of ccnd.