Removed a number of InterruptedExceptions and their aftermath.
Re-fixed CCNInputStream. That code is designed to read arbitrary block streams, where the only guarantee is that the last (pre-digest) component of the block name is a monotonically increasing value (treated as an integer), though those do not need to be consecutive. While this works to
read block-fragmented data with standard fragment numbers, it is also designed to read other
forms of streaming data (e.g. rtp streams generated by VoCCN) and so cannot assume that the
content it reads will have fragmentation markers in its name (or version markers), or that it will have a header block. (In addition, it is also meant to read content while it is still being streamed -- e.g. before its header block has been created -- and only pick up the header block asynchronously.)