This commit addresses refs #26, #27, #220, #219, #218
Issues #26 and #27 should be resolved, as they now publish a file to a
repository from the file system using the GUI, and then read a file back
to file system. Currently the contents of the repo and name enumerator
are all synthetic, that is they are hard coded into a class called
GUINameEnumerator. Also, the only file types it expects currently are
txt files, and the contents of the txt file captured in the UI are shown
in the lower panel.

Additionally, to see this correctly on Ubuntu you have to set the
enviroment variable AWT_Toolkit=MToolkit.

Added icon files for directory and file icons