Move repo interest handling to separate class. Just wait for exit request in work() refs #309
Fix various timing problems related to making the repo do all its work in handlers.

This includes the problem of trying to change the filter due to a namespace
change while there are still outstanding sessions and timing problems on the
client side - namely that in some cases it could cancel an ack interest before
the dataresponding to that interest had been received and that the immediate
ack request after close was sometimes received before the repo actually had
all the data so the repo could not respond with all the data it had at that
time.  refs #309