Fix several aspects of the c build:
CFLAGS may now be used from the top-level make to control things like
the optimization level and debugging symbols, without disrupting the
other options required by the build process.

It now works to cd into a subdirectory of src/ and type 'make'.

Placing the build products in a subdirectory is still possible (make objtree).
The default directory name for this is just "build" now, use OBJTREE=...
to provide your own.

refs #428  Reorganize c sources into a more sensible hierarchy