Implement multicast faces in ccnd
At long last.

Code in ccn_sockcreate.c largely came from udplink, with modifications for
the way errors are reported and handled.  It needs re-indenting.

The bind-to-specific-interface case is coded, but not yet tested.

There is not yet a timeout on the multicast or stream faces, nor a way to destroy them.

For now there is some unconditional logging.

refs #89  udplink merged into ccnd