Fixed a couple of logic bugs in CCNFlowControl. 1) Would buffer a content object that had a pending interest match and was sent out immediately, potentially leaving content in buffer that would never be requested again because expected interest had already been satisfied. 2) Multiple writer threads (if that were to happen) could all add objects to buffer then wait, pushing the buffer content well above highwater. This probably would not occur in practice, since multiple threads really should not be writing, but it is clearer to verify the algorithm if threads block on full buffer before adding content to the data structure.
Note that this change alters the semantics of the highwater mark
slightly: the buffer will now accept up to highwatermark objects without
blocking the caller.  So if (as in the unit test) you call setHighwater(4)
you can expect to insert 4 objects without blocking even when there is
no reader to consume them.

Refs #738