Refactor ccnd for improved flexibility in file locations and logging.
It may not always be desirable to produce ccnd's messages on stderr;
change ccnd_msg so the log messages may be handled by whatever means are
desired. At present, actually taking advantage of this will require using
a different main procedure, so pull that out into its own source file and
remove calls to exit() outside of the main.

Add environment variables to override the defaults for the unix-domain socket
used to contact ccnd, and for the directory used for ccnd's keystore.

refs #100050, #100048, #100031