Make a more palatable c API for signing content objects.
Introduce ccn_sign_content(h, ...), which gets help from the library
handle to help manage signing keys, and has a better API for specifying
the parameters.

The older API is still there (and is used underneath the new one), but so
far only the ccnput command has been converted.

Note that the template portion of the API is not yet implemented.

CCN_API_VERSION bumped to 1004

refs #100072
refs #100038