Revert "Bug fix in AccessControlManager"
This reverts commit b0b0e8d54ced02d8af60ee325fcc502fdeebaf05.

This turned out not to be a bug; we changed it and then fixed what
was probably the actual bug (759aec54f97e3b2cdfdbb033e54c6d707637c661),
and for some strange reason it worked in our test case. Don't know
why... Back it out and see what happens.