Reorganize how ccnd keeps track of the outbound faces for an interest.
In the new representation, pe->outbound is kept in send order instead
of the reverse, and the new index pe->sent is advanced as the faces
are used.  This leaves a trail of the faces that have been used already,
giving more flexibility in altering the outbound list to the right
of pe->sent.

refs #100103