Simplify seqwriter's internal state.
It now only attempts to produce one block at a time.  The first
block is sent even in the absence of interest, but a copy is kept
so that it can be resent in case ccnd drops it.  Other blocks obey
strict flow balance.

Note that the first block may well be a short one because of this strategy.