Add facility based logging.  Facilities have independent log levels. They can be set via a system property or an environment variable (as per SystemConfiguration).
The current facilities are: Default, Pipelining, and NetManager.  Only
Pipelining is currently using a facility.

The default log levels are INFO for Default and NetManager and WARNING
for Pipelining.

The system properties are org.ccnx.ccn.LogLevel (for default) and
.Pipeline and .NetManager for the others.

The environment variables are CCN_LOG_LEVEL (for default) and _PIPELINE
and _NETMANAGER for the others.

There is now a junit test for logging (test/impl/LogTest) that tests the
basics for facility based logging.