Check the ccnd face representation so that it uses a faceid instead of a socket for keepin track of where to send things.
This means we don't have to worry about accidentally closing the wrong
socket, since faceids are not reused.

The multicast case is a little broken by this - still needs work.
Also, recv_fd is now a counfusing name and it will be changed;
keeping it for now to ease merges.