Introduce new CCN_FORW_LOCAL forwarding flag.
The presence of this flag in a FIB entry marks the namespace as local
to the host (more specfically, to apps directly connected to the ccnd
in question).  This is in support of implementing a host-local namespace
useful for local service discovery. refs #100134

The ccnd does its best to make sure that content within such a namespace
comes only from local apps, and is delivered only to local apps.

There is one known flaw in the current implementation, namely that
an off-machine attacker could cause a stale content object in the
local space to become fresh again.  To do this, the attacker would
have to already have a copy of the C.O., which was supposed to be
machine-local. Thus there must also be some collusion going on for
this to happen.