Correct opening of stream with segment number present in content name
If the initial segment desired for a stream is specified by passing
it as part of the name being opened, it must be removed to form the
base name.

Calls to the constructor
   CCNAbstractInputStream(ContentName n, Long startingSegmentNumber, ...)
which specify both a segment number in the content name and an explicit
starting segment number will result in the explicit starting segment number
taking precedence.

Improve logging in case where the base name of a stream does not match
the name of the current segment.