Correct Logging unit test failure   Refs #100185
The DEFAULT_APPLICATION_CLASS must be a prefix in the Logging namespace
of the other loggers in order that they inherit the handlers added to it.
We can reduce the number of handlers added to the different loggers by
making use of that fact, and enable applications (or tests) to add specialized
handlers to the default application class logger without requiring that they
do so for each facility logger.