refs #100226 The presence of a forward_to entry should not prevent ccnd from pruning a prefix.
Fom some time now, forward_to is just cached information, and so it it
fine to remove a prefix entry that has one, as long as the other
conditions for removal are satisfied.

Otherwise the entry is not removed until all of the faces involved have
gone away, which can be a long time.

It would probably be better to create the 'tracking' prefix entries at
one level higher in the tree, which would cut down the churn quite a bit.