refs #100275  Add sequence number generation
In this version, 16-bit sequence numbers are used, starting at a random
place and allowed to wrap.  Logging is done on the sending side, but this
will probably be removed once things seem to be working.

For compatibility, an experimental sequence number is produced at the
outset, but no more are produced after that unless the other side generates them
as well.  This will cause just one log message on the receiving side if it happens
to be a version that does not understand sequence numbers.