refs #100461. Stuff an interest if we have received nothing for a while.
We ask for something we know should be available (a ccnd key in the neighborhood namespace),
but we don't care very much about the answer; we just want to elicit traffic from the
other side to see if the link is up.

By doing this as a stuffed interest, we don't cause extra interest traffic (in terms of packets).

Right now this is for the unicast datagram case only.