Introduce new ccnr_accession and ccnr_hwm data types
These two new data types and their operations should be used in place
of the old ccn_accession_t.  Places in the code that are likely to
need attention have been marked with XXXXXX.  Declarations that use
ccnr_accession should also be reviewed and switched to ccnr_hwm
where appropriate.

Compiling with -DCCNLINT set to 1, 2, or 3 will help flag things that
need changing.