refs #100626 Add missing android.mk file for lib2, fix build to complete at controller module.
Problem with current dependency between ccnr and sync, it doesn't make sense to implement the build in the same manner as the ccnd module.  It's dependencies are contained within the module.  ccnr and sync have some cross dependency.  It makes better sense to compile the lib2 objects and sync objects into libccnr.  May need to modify this to support standalone sync in the future.  So I've left the Android.mk files in for sync and lib2.