refs #100626 Add check for existence of repo_dir and create it if needed, and handle bailout properly.
Was not checking for existence of repo_dir.  Updated repo1 and repo2 to use same pathway for this.  Take out hardcoded path and move up to 'const' definition in the DEFAULTS.  If we hit a problem while creating the repo_dir, we should bail out.  Handle this properly for repo2 case.  Add some better logging.  Set the options to also set the proto default = 'unix'.  None of this is configurable at the moment, other than by overriding the implementation here.  Will add a the toggle to the CCNx Android Services to use repo1 or repo2.  Should also set sync on toggle.  At the moment sync is set to 0 (off).  Expect to update this again when default CCNR env variables are changes to match consistent naming scheme (if this happens) before v0.5 release.  Current state is that chat works.