refs #100626 Per Nick's review, clean up pathway thru onStartService to handle v1/v2 properly.
Split out env keys in RepoWrapper for SYNC_OPTIONS, CCNR_OPTIONS, and REPO_OPTIONS.  This can be cleaned up too after we finalize new env var naming.  Make sure we have the defaults covered for the variables that require some android specific configuration.  Otherwise, we now read in the config (if any) found in the onStartService, set these into options, and then options are set via ccnrSetenv.   If not set, then we fallback to defaults for those env variables.