refs #100626 Implement very basic UI for setting properties for CCNR.
Added a sort of form interface for handling set values for CCNR_DIRECTORY, CCNR_GLOBAL_PREFIX, and CCNR_DEBUG.  All other properties are to be set by the config or env.  Needs a bit more cleanup.  Basic services work.  We need more extensive test cases for the service to make sure it works well, or works well in a "long running" state (leave it one for a weekend and see if the service is still running).  Next commit will refactor in changes brought in by #100638.