Refs #100726 Additional flush of block FIFO when seeking.
It is necessary to flush the block FIFO before taking the mutex in
order unblock the incoming content handler when it is blocked in
block_FifoPace() while holding the mutex.  It may then place another
block in the FIFO before giving up the lock, so Seek must re-flush the
FIFO after obtaining the lock.