Refs #100755  Keep a scope 0 interest around, if it is not satisfied from the cache.
The matching content may arrive later.  This enables a way of collecting
up content as it arrives at the local ccnd, by an app that echos back
the interests it receives, only with scope 0.

There are a few situations where this does not work as desired.  One is if
the content arrives based on a less-specific interest than the prefix
that the snooping app has registered.