refs #100645 Add additional debugging, set error status on exceptions, and implement button disable when start/stop clicked.
Avoids many of the service startup/shutdown issues seen with multiple clicks.
Still need a manual button to just reset the entire service to stopped
service end/finished.  I think this can be added to a menu item.  This
will save the user a trip to the Applications manager to manually
shutdown the service.  Needs testing in case that System load library
fails.  Previously this would keep failing and make the device very
unusable with android service failures.