refs #100645 Handle worst case of CCNxService startup failure.
This shouldn't ever happen.  However, it can and does.  When it does,
the CCNxService crashes off in the background, sometimes repeatedly.
The right thing to do is indetify these error states and make sure we
1. Log it
2. Report it in the UI
3. Do any cleanup needed so that next run attempt doesn't trip over
leftover artifacts (dot files, etc.)
4. Rethrow any exceptions we can manage to catch
5. Shutdown nicely, don't keep on trying to run.

Clearly #1 is going to get covered by all of the stacktraces in the
code.  However, that doesn't always yield enough context to someone
staring at the logs as to why.  So if there is some additional info,
log it, and explain what is going to happen next.  #2 we now at
least have enough error status markers in the code to set the state
properly when exceptions get swallowed.  And it's not just a toast,
so we keep error status in place until it's manually cleared by the
user.  #3, we do delete the ccnd keystore remenants if it fails to
create the keystore due to BCP not found.  #4, instead of hitting a
Runtime Error with NoClassDefFoundError, added CCNxLibraryCheck to
test for existence of BCP in the classloader.  If it's there, no
exception.  If it isn't there, we don't pass on an error, but throw
an exception instead.  Yes, this is frowned upon Practice, but being
pratical, the alternative is the service crashes.  Not what we want.
And #5 Make an attempt to stop anything that was running, and don't skip
over the exceptions like they never happened.  And don't ignore the
error states.  The only way to transition to SERVICE_OFF is to get
to SERVICE_FINISHED.