refs #100321 Major revision to the way ccnd manages pending interests
The hash table of interest nonces is gone.  Instead there is a hash
table of interest bodies (sans nonces), and the per-face info
(including associated nonces) hangs off of that.

This version compiles, but does not work correctly.  It still needs
to be taught how to put the info about outbounds into the pit.