Refs #100851  Use an outer loop with the test at the bottom.
It turns out that this produces no significant difference in the number of
gettime calls (over the previous commit), though it could in theory.
It is attractive to have only one update_time call in the procedure, so
leave it this way.

There are still about 15-20% more calls to update_time than there
were before this set of chenges.  This does not seem an excessive price
to pay for correct behavior.