Refs #100782 fix return value of ccn_resolve_version() to behave as documented.
Required followup fix to ccn_get_header() to match documented behavior rather
than observed behavior.
Documented behavior is that ccn_resolve_version() will return 0 if the name has
not been extended, or 1 if it has been extended.  Note that 0 does not mean that
the name is not versioned -- it may have had a pre-existing version.