Refs #100898  Fix use-after-free problem
The reason this was happening is that two events could get
scheduled for the same ie, but there is only one slot to
keep track of it.  The fix cancels the interlopers.  There
may be a better way to deal with this, but it ought to be
a rare event.