Refs #100890, 100118, 100119  Change the way ccnd tracks stale data
Instead of former mark-and-reap method, keep track of the order in
which the content objects become stale.  This means it is always
easy to find the next thing to remove, so just do that when the
capacity has been reached.

For now, this uses a doubly-linked list.  This will later be augmented
to prevent the insertion cost from getting too high.

The schedule mechanism is no longer needed for staleness.

Since nothing happens at the time that content becomes stale, there
is no longer a log entry made at that time.  Removal is still loggable.

With this change, content objects with FreshnessSeconds of 0 are
allowed (bug #100118). Large values are also allowed (bug #100119);
there is a somewhat arbitrary cap of 4 years.