Refs #100906  Reintroduce the table of interest nonces, for dropping duplicates.
Interest logging will show interest_dupnonce when an interest is dropped
due to nonce duplication.  The nonces time out after 6 seconds from their
most recent reception (this should probably be made tunable).

The status page shows the size of the nonce table as the "noted" count.
The "propagating" count is really the size of the interest table.