refs #100911 Added 'make pkgbin' target to top level makefile to make a binary release
Conceptually this is similar to 'make dist' distfile, in that it will
package up the release into a tarball.  Where it is different is that it
does not package source, and is based on the regular 'make' default
target, which after building, uses 'make install' to install into a
directory: ccnx-pkg-<VERSION> , and creates a tarball based on this.

The layout inside of ccnx-pkg-<VERSION> is:
- bin
- include
- lib
- man
- doc
- etc

We don't yet use etc, but it should be expected that we'll put some
default config here, as well as daemon scripts when we have created what
we need.  Manpages are moved from the underlying doc directory into the
man page dir.  Everything else is moved according to 'make install' with
a combination of DESTDIR and INSTALL_BASE.