Refs #100984  Remove the hash table for strategy instances.  Add some comments.
Maintaining the hash table so that indices could be used instead of a direct
pointer turned out to take quite a lot of extra work for little or no space
savings.  Performance probably would have been worse as well, though this
was not measured.  We are now walking up the tree to find the active strategy;
this could be cached, but it is not at all clear it would be worth it, so keep
it simple for now.