Refs #100983, 100984 Updates to the strategy API.
Instead of calling the strategy multiple times for new upstreams,
batch the requests into one CCNST_UPDATE callout, calling attention
to the upstreams that require a strategy decision to send or not.

The default behavior is changed so that sending never happens without
the strategy say-so (except for TAP prefixes).  A viable strategy
must do some processing in the CCNST_UPDATE case.