Refs #101044  Add a way to insert limits for test ccnd instances.
This adds a way to hook in ulimit commands that are to be executed
before a capacity-limited ccnd is started.

Since the details of the ulimit command differ from system to system,
not to mention the memory requirements, attempting a portable check for
unexpected heap growth is not worth it.  This hook will allow test
systems to be set up that look for such problems.