Refs #101049  Add a per-cob refcount to know how many face queues it is in.
This permits us to try to avoid discarding content that is about to be sent.

There would be a DOS risk if a face accumulates a lot of queued entries,
so if there are too many referenced items at the head of the expiry line,
go ahead and take the first one anyway.