Refs #101060  Stop examining the cache if we have reached the configured cache limit.
This should be acceptable in theory, since we match against the pit upon
receipt of a duplicate content object.  The external behavior is
indistinguishable from that which would have resulted if the content
object had been ejected from the cache.  In practice, some tests
rely on the presence of the cache, so for now the default
CCND_MAX_MATCH_PROBES is larger than it ideally should be to
accomodate these tests until they can be fixed or removed.